==> 00-preamble.R <==

# load packages -----------------------------------------------------------
library(readr)
library(tidyverse)
library(paletteer)
library(hrbrthemes)
library(dggridR)
library(sf)
library(lme4)
library(stringr)
library(maps)
library(emmeans)
library(showtext)
library(see)
library(patchwork)
font_add_google("Roboto Condensed") #just run once
cc <- paletteer_d("ggthemes::calc")
